requirements.txt
tensorflow==2.7.0
torch==1.10.1
botorch==0.6.0
numpy==1.20.3
matplotlib==3.5.1
scikit-learn>=1.0.2
xgboost==1.5.1
scipy==1.7.1
Cython==0.29.23
ConfigSpace==0.4.18
hyperopt==0.2.7
pyyaml==5.4.1